chore: update error message for column type
